requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.5
python-dotenv>=1.0.0
requests>=2.31.0
PyYAML>=6.0
beautifulsoup4>=4.12.2
SQLAlchemy>=2.0.20
SQLAlchemy-Utils>=0.41.1
schedule>=1.2.0
colorlog>=6.7.0
aiosqlite>=0.19.0
APScheduler>=3.10.4
pydantic>=2.3.0
discord-ui>=5.1.6
discord-py-interactions>=4.4.1
tabulate>=0.9.0
python-dateutil>=2.8.2 